Implement loading of the vst structure from a json file

+++ runtime/hetdesrun/adapters/virtual_structure_adapter/utils.py
@@ -2,12 +2,13 @@ import logging
 from uuid import UUID
 
 from hetdesrun.models.wiring import InputWiring, OutputWiring
-from hetdesrun.structure.models import Sink, Source
+from hetdesrun.structure.models import CompleteStructure, Sink, Source
 from hetdesrun.structure.structure_service import (
     delete_structure,
     get_collection_of_sinks_from_db,
     get_collection_of_sources_from_db,
     is_database_empty,
+    load_structure_from_json_file,
     update_structure,
 )
 from hetdesrun.webservice.config import get_config
@@ -69,30 +70,48 @@ def prepopulate_structure() -> None:
         logger.info(
             "Structure of the virtual structure adapter was not prepopulated, "
             "because the environment variable "
-            "'PREPOPULATE_VST_ADAPTER_AT_HD_STARTUP' is set to false"
+            "'PREPOPULATE_VST_ADAPTER_AT_HD_STARTUP' is set to False"
         )
         return
+
+    complete_structure = retrieve_complete_structure_for_prepopulation()
+
+    if (
+        get_config().completely_overwrite_an_existing_virtual_structure_at_hd_startup
+        and not is_database_empty()
+    ):
+        delete_structure()
+        logger.info(
+            "An existing structure was found in the database and deleted, "
+            "during the prepopulation process of the virtual structure adapter"
+        )
+    update_structure(complete_structure)
+    logger.info("The structure was successfully updated.")
+
+
+def retrieve_complete_structure_for_prepopulation() -> CompleteStructure:
+    if get_config().prepopulate_virtual_structure_adapter_via_file:
+        logger.info("Prepopulating the virtual structure adapter via a JSON-file")
+        if (
+            structure_filepath
+            := get_config().structure_filepath_to_prepopulate_virtual_structure_adapter
+        ):
+            return load_structure_from_json_file(structure_filepath)
+        raise ValueError(
+            "If prepopulation of the virtual structure adapter structure "
+            "via a file is set, "
+            "'STRUCTURE_FILEPATH_TO_PREPOPULATE_VST_ADAPTER' must be set, "
+            "but it is not."
+        )
+
     if complete_structure := get_config().structure_to_prepopulate_virtual_structure_adapter:
-        is_db_empty = is_database_empty()
-        if get_config().completely_overwrite_an_existing_virtual_structure_at_hd_startup:
-            if not is_db_empty:
-                delete_structure()
-                logger.info(
-                    "An existing structure was found in the database and deleted, "
-                    "during the prepopulation process of the virtual structure adapter"
-                )
-            update_structure(complete_structure)
-        else:
-            update_structure(complete_structure)
-            logger.info(
-                "%s and the update_structure function was executed.",
-                "An existing structure was found in the database"
-                if is_db_empty
-                else "No existing structure was found in the database",
-            )
-        return
-    logger.info(
-        "Structure of the virtual structure adapter was not prepopulated, "
-        "because no structure was provided "
-        "via the environment variable 'STRUCTURE_TO_PREPOPULATE_VST_ADAPTER'"
+        logger.info(
+            "Prepopulating the virtual structure adapter via the environment variable "
+            "'STRUCTURE_TO_PREPOPULATE_VST_ADAPTER'"
+        )
+        return complete_structure
+
+    raise ValueError(
+        "The prepopulation of the virtual structure adapter structure "
+        "is enabled, but no structure was provided."
     )
